用户浏览记录日志

+++ Picture/picture-picture/src/main/java/com/lz/picture/model/domain/UserViewLogInfo.java
@@ -4,6 +4,7 @@ import java.io.Serializable;
 import java.util.Map;
 import java.util.Date;
 import java.math.BigDecimal;
+
 import com.baomidou.mybatisplus.annotation.IdType;
 import com.baomidou.mybatisplus.annotation.TableId;
 import com.baomidou.mybatisplus.annotation.TableField;
@@ -12,6 +13,7 @@ import lombok.Data;
 import com.lz.common.annotation.Excel;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonInclude;
+
 /**
  * 用户浏览记录对象 p_user_view_log_info
  *
@@ -20,77 +22,110 @@ import com.fasterxml.jackson.annotation.JsonInclude;
  */
 @TableName("p_user_view_log_info")
 @Data
-public class UserViewLogInfo implements Serializable
-{
+public class UserViewLogInfo implements Serializable {
     private static final long serialVersionUID = 1L;
 
-    /** 记录编号 */
-        @Excel(name = "记录编号")
+    /**
+     * 记录编号
+     */
+    @Excel(name = "记录编号")
     @TableId(value = "view_id", type = IdType.ASSIGN_ID)
     private String viewId;
 
-    /** 用户编号 */
-        @Excel(name = "用户编号")
+    /**
+     * 用户编号
+     */
+    @Excel(name = "用户编号")
     private String userId;
 
-    /** 目标类型 */
-        @Excel(name = "目标类型")
+    /**
+     * 目标类型
+     */
+    @Excel(name = "目标类型")
     private String targetType;
 
-    /** 目标对象 */
-        @Excel(name = "目标对象")
+    /**
+     * 目标对象
+     */
+    @Excel(name = "目标对象")
     private String targetId;
 
-    /** 目标内容 */
-        @Excel(name = "目标内容")
+    /**
+     * 目标内容
+     */
+    @Excel(name = "目标内容")
     private String targetContent;
 
-    /** 分数 */
-        @Excel(name = "分数")
-    private BigDecimal score;
+    /**
+     * 分数
+     */
+    @Excel(name = "分数")
+    private Double score;
 
-    /** 图片分类 */
-        @Excel(name = "图片分类")
+    /**
+     * 图片分类
+     */
+    @Excel(name = "图片分类")
     private String categoryId;
 
-    /** 空间 */
-        @Excel(name = "空间")
+    /**
+     * 空间
+     */
+    @Excel(name = "空间")
     private String spaceId;
 
-    /** 图片标签 */
-        @Excel(name = "图片标签")
+    /**
+     * 图片标签
+     */
+    @Excel(name = "图片标签")
     private String tags;
 
-    /** 封面 */
-        @Excel(name = "封面")
+    /**
+     * 封面
+     */
+    @Excel(name = "封面")
     private String targetCover;
 
-    /** 查看时间 */
-        @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
-        @Excel(name = "查看时间", width = 30, dateFormat = "yyyy-MM-dd HH:mm:ss")
+    /**
+     * 查看时间
+     */
+    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
+    @Excel(name = "查看时间", width = 30, dateFormat = "yyyy-MM-dd HH:mm:ss")
     private Date createTime;
 
-    /** 设备唯一标识 */
-        @Excel(name = "设备唯一标识")
+    /**
+     * 设备唯一标识
+     */
+    @Excel(name = "设备唯一标识")
     private String deviceId;
 
-    /** 浏览器类型 */
-        @Excel(name = "浏览器类型")
+    /**
+     * 浏览器类型
+     */
+    @Excel(name = "浏览器类型")
     private String browser;
 
-    /** 操作系统 */
-        @Excel(name = "操作系统")
+    /**
+     * 操作系统
+     */
+    @Excel(name = "操作系统")
     private String os;
 
-    /** 平台 */
-        @Excel(name = "平台")
+    /**
+     * 平台
+     */
+    @Excel(name = "平台")
     private String platform;
 
-    /** IP属地 */
-        @Excel(name = "IP属地")
+    /**
+     * IP属地
+     */
+    @Excel(name = "IP属地")
     private String ipAddress;
 
-    /** 请求参数 */
+    /**
+     * 请求参数
+     */
     @JsonInclude(JsonInclude.Include.NON_EMPTY)
     @TableField(exist = false)
     private Map<String, Object> params;
